+++ lib/view_component/base.rb
@@ -54,6 +54,7 @@ module ViewComponent
     include ViewComponent::Translatable
     include ViewComponent::WithContentHelper
 
+    # Config option that strips trailing whitespace in templates before compiling them.
     class_attribute :__vc_strip_trailing_whitespace, instance_accessor: false, instance_predicate: false, default: false
 
     # For CSRF authenticity tokens in forms
@@ -67,8 +68,6 @@ module ViewComponent
 
     # For Content Security Policy nonces
     delegate :content_security_policy_nonce, to: :helpers
-
-    # Config option that strips trailing whitespace in templates before compiling them.
 
     attr_accessor :__vc_original_view_context
     attr_reader :current_template