+++ test/sandbox/app/components/no_cache_component.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class NoCacheComponent < ViewComponent::Base
-  include ViewComponent::Cacheable
+  include ViewComponent::ExperimentallyCacheable
 
   attr_reader :foo, :bar